Update GPL license links to use HTTPS
Replace all occurrences of the GNU General Public License HTTP link with its HTTPS counterpart for improved security and consistency across the project files.
