+++ src/main/java/de/greluc/sc/sckm/ScKillMonitorApp.java
@@ -15,7 +15,7 @@
  * GNU General Public License for more details.                                                   *
  *                                                                                                *
  * You should have received a copy of the GNU General Public License                              *
- * along with SC Kill Monitor. If not, see <http://www.gnu.org/licenses/>.                        *
+ * along with SC Kill Monitor. If not, see <https://www.gnu.org/licenses/>.                       *
  **************************************************************************************************/
 
 package de.greluc.sc.sckillmonitor;
