+++ src/test/java/de/greluc/sc/sckm/settings/SettingsDataTest.java
@@ -15,12 +15,63 @@
  * GNU General Public License for more details.                                                   *
  *                                                                                                *
  * You should have received a copy of the GNU General Public License                              *
- * along with SC Kill Monitor. If not, see <http://www.gnu.org/licenses/>.                        *
+ * along with SC Kill Monitor. If not, see <https://www.gnu.org/licenses/>.                       *
  **************************************************************************************************/
+
+package de.greluc.sc.sckillmonitor.settings;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.*;
 
 /**
  * @author Lucas Greuloch (greluc, lucas.greuloch@protonmail.com)
  * @since 1.0.0
  * @version 1.0.0
  */
-package de.greluc.sc.sckillmonitor.settings;
+class SettingsDataTest {
+
+  @Test
+  void testSetSelectedChannelUpdatesValue() {
+    // Arrange
+    String newChannel = "PTU";
+
+    // Act
+    SettingsData.setSelectedChannel(newChannel);
+
+    // Assert
+    assertEquals(newChannel, SettingsData.getSelectedChannel());
+  }
+
+  @Test
+  void testSetSelectedChannelNotifiesListeners() {
+    // Arrange
+    String newChannel = "TECH-PREVIEW";
+    SettingsListener mockListener = mock(SettingsListener.class);
+    SettingsData.addListener(mockListener);
+
+    // Act
+    SettingsData.setSelectedChannel(newChannel);
+
+    // Assert
+    verify(mockListener, times(1)).settingsChanged();
+
+    // Cleanup
+    SettingsData.removeListener(mockListener);
+  }
+
+  @Test
+  void testSetSelectedChannelDoesNotNotifyRemovedListener() {
+    // Arrange
+    String newChannel = "HOTFIX";
+    SettingsListener mockListener = mock(SettingsListener.class);
+    SettingsData.addListener(mockListener);
+    SettingsData.removeListener(mockListener);
+
+    // Act
+    SettingsData.setSelectedChannel(newChannel);
+
+    // Assert
+    verify(mockListener, never()).settingsChanged();
+  }
+}
